+++ lib/main.js
@@ -18,22 +18,16 @@
 // },
 //
 
-import {Emitter, CompositeDisposable, Disposable, Range, Point, TextEditor} from "atom"
+import {Emitter, CompositeDisposable, Disposable} from "atom"
 
-// import _ from "lodash"
-// import {autorun} from "mobx"
 import React from "react"
 import ReactDOM from "react-dom"
 
 import KernelPicker from "./kernel-picker"
 import WSKernelPicker from "./ws-kernel-picker"
 import SignalListView from "./signal-list-view"
-import * as codeManager from "./code-manager"
 
-import Inspector from "./components/inspector"
-import ResultView from "./components/result-view"
 import StatusBar from "./components/status-bar"
-
 import InspectorPane from "./panes/inspector"
 import WatchesPane from "./panes/watches"
 import OutputPane from "./panes/output-area"
@@ -41,25 +35,19 @@ import OutputPane from "./panes/output-area"
 import {toggleInspector} from "./commands"
 
 import store from "./store"
-import OutputStore from "./store/output"
 
 import Config from "./config"
 import kernelManager from "./kernel-manager"
 import bubbleManager from "./bubble-manager"
 import ZMQKernel from "./zmq-kernel"
-import WSKernel from "./ws-kernel"
 import AutocompleteProvider from "./autocomplete-provider"
 import HydrogenProvider from "./plugin-api/hydrogen-provider"
 import {
-  log,
-  reactFactory,
-  isMultilanguageGrammar,
   renderMobxReactDevtools,
   INSPECTOR_URI,
   WATCHES_URI,
   OUTPUT_AREA_URI,
   hotReloadPackage,
-  getGrammar,
 } from "./utils"
 
 import Hydrogen from "./hydrogen"
@@ -68,7 +56,7 @@ function getEditor() {
   return atom.workspace.getActiveTextEditor()
 }
 
-const HydrogenPackage = {
+export default {
   config: Config.schema,
 
   activate() {
@@ -187,12 +175,14 @@ const HydrogenPackage = {
     this.subscriptions.add(new Disposable(() => tile.destroy()))
   },
 
+  // Section: Control Kernel
+  //==================================================
   // FIXME: BROKEN
   switchKernel(kernelSpec) {
     if (!kernelSpec) return
-    bubbleManager.clear(getEditor())
-
     const editor = getEditor()
+    bubbleManager.clear(editor)
+
     const kernel = kernelManager.getKernelForEditor(editor)
     if (kernel) kernel.destroy()
 
@@ -226,120 +216,30 @@ const HydrogenPackage = {
     this.getHydrogen().restartKernelAndReEvaluateBubbles()
   },
 
+  // Section: Delegate to Hydrogen instance
+  //==================================================
   toggleBubble() {
-    const editor = getEditor()
-    const [startRow, endRow] = editor.getLastSelection().getBufferRowRange()
-
-    const markerLayer = bubbleManager.getMarkerLayer(editor)
-
-    for (let row = startRow; row <= endRow; row++) {
-      let destroyed = false
-
-      for (const marker of markerLayer.findMarkers({startBufferRow: row})) {
-        marker.destroy()
-        destroyed = true
-      }
-
-      if (!destroyed) {
-        const outputStore = this.createOutputStore(editor, row)
-        outputStore.status = "empty"
-
-        bubbleManager.createBubble({
-          editor: editor,
-          row: row,
-          view: new ResultView(outputStore, null, true),
-        })
-      }
-    }
+    this.getHydrogen().toggleBubble()
   },
 
-  run({moveDown = false} = {}) {
-    const editor = getEditor()
-    const codeBlock = codeManager.findCodeBlock(editor)
-    if (!codeBlock) return
-    const [text, row] = codeBlock
-
-    if (!text) return
-    if (moveDown) codeManager.moveDown(editor, row)
-
-    this.executeText(editor, text, row)
+  run(...args) {
+    this.getHydrogen().run(...args)
   },
 
   runAll(breakpoints) {
     this.getHydrogen().runAll(breakpoints)
   },
 
-  createOutputStore(editor, row) {
-    const outputStore = new OutputStore()
-    outputStore.updatePosition({
-      lineLength: editor.buffer.lineLengthForRow(row),
-      lineHeight: editor.getLineHeightInPixels(),
-      editorWidth: editor.getEditorWidthInChars(),
-    })
-    return outputStore
-  },
-
-  async executeText(editor, text, bubbleRow) {
-    await kernelManager.startKernelIfNeeded(editor)
-    const kernel = kernelManager.getKernelForEditor(editor)
-
-    const outputStore = this.createOutputStore(editor, bubbleRow)
-    const decoration = bubbleManager.createBubble({
-      editor: editor,
-      row: bubbleRow,
-      view: new ResultView(outputStore, kernel, true),
-    })
-
-    kernel.execute(text, message => {
-      editor.component.invalidateBlockDecorationDimensions(decoration)
-      outputStore.appendOutput(message)
-    })
-  },
-
   runAllAbove() {
-    const editor = getEditor()
-    bubbleManager.clear(editor)
-
-    if (isMultilanguageGrammar(editor.getGrammar())) {
-      atom.notifications.addError('"Run All Above" is not supported for this file type!')
-      return
-    }
-
-    const cursor = editor.getLastCursor()
-    const row = codeManager.escapeBlankRows(editor, 0, cursor.getBufferRow())
-    const text = codeManager.normalizeString(editor.getTextInBufferRange([[0, 0], [row, Infinity]]))
-    if (!text) return
-
-    this.executeText(editor, text, row)
+    this.getHydrogen().runAllAbove()
   },
 
-  runCell({moveDown = false} = {}) {
-    const editor = getEditor()
-
-    const range = codeManager.getCurrentCell(editor)
-
-    const {start, end} = range
-
-    // FIXME: recheck
-    const _range = new Range(start, [end.row, Infinity])
-
-    for (const marker of bubbleManager.getMarkers(editor)) {
-      if (_range.containsPoint(marker.getStartBufferPosition())) {
-        marker.destroy()
-      }
-    }
-
-    const endRow = codeManager.escapeBlankRows(editor, start.row, end.row)
-    const text = codeManager.normalizeString(
-      editor.getTextInBufferRange([[start.row, 0], [endRow, Infinity]])
-    )
-    if (!text) return
-
-    if (moveDown) codeManager.moveDown(editor, endRow)
-
-    this.executeText(editor, text, endRow)
+  runCell(...args) {
+    this.getHydrogen().runCell(...args)
   },
 
+  // Section: Ignore these in the meanwhile.
+  //==================================================
   // TODO
   provideHydrogen() {
     if (!this.hydrogenProvider) {
@@ -391,5 +291,3 @@ const HydrogenPackage = {
     )
   },
 }
-
-export default HydrogenPackage
